Fix test time value and add missing blacklist tests
The blacklist functionality on master didn't have the updates on develop, meaning we couldn't actually blacklist expired tokens (despite the fact that they could be refreshed). First we'll create tests for this, and also remove `time()` calls from other tests, preferring to use a fixed time based on `Carbon::setTestNow`.